# MQE: Add NH support to binary operations