… CHANGELOG.md
… pkg/streamingpromql/engine_test.go
… pkg/streamingpromql/operators/binary_operation_test.go
… pkg/streamingpromql/testdata/ours/binary_operators.test
… pkg/streamingpromql/types/data.go
… pkg/streamingpromql/types/data_test.go