@jhesketh 2024-07-29
WIP, no tests yet.

Not the most elegant. The merging one side and many-to-one mapping may need rethinking.