@jhesketh 2024-08-06
This branch vs promql engine:

<details>

```
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/streamingpromql/benchmarks
cpu: Intel(R) Xeon(R) CPU E5-2686 v4 @ 2.30GHz
                                                                                                         │      Prometheus      │                Mimir                 │
                                                                                                         │        sec/op        │    sec/op     vs base                │
Query/a_1,_instant_query                                                                                            674.1µ ± 2%   659.7µ ±  3%        ~ (p=0.093 n=6)
Query/a_1,_range_query_with_100_steps                                                                               702.2µ ± 2%   688.0µ ±  4%   -2.02% (p=0.004 n=6)
Query/a_1,_range_query_with_1000_steps                                                                              863.4µ ± 3%   852.9µ ±  2%        ~ (p=0.310 n=6)
Query/a_100,_instant_query                                                                                          1.863m ± 2%   1.843m ±  1%   -1.10% (p=0.041 n=6)
Query/a_100,_range_query_with_100_steps                                                                             2.868m ± 1%   2.831m ±  1%   -1.26% (p=0.026 n=6)
Query/a_100,_range_query_with_1000_steps                                                                            12.56m ± 2%   12.51m ±  2%        ~ (p=0.485 n=6)
Query/a_2000,_instant_query                                                                                         21.66m ± 1%   21.36m ±  1%   -1.38% (p=0.004 n=6)
Query/a_2000,_range_query_with_100_steps                                                                            42.08m ± 1%   41.61m ±  2%        ~ (p=0.065 n=6)
Query/a_2000,_range_query_with_1000_steps                                                                           233.0m ± 1%   227.2m ±  1%   -2.45% (p=0.002 n=6)
Query/nh_1,_instant_query                                                                                           762.4µ ± 5%   738.6µ ±  4%        ~ (p=0.240 n=6)
Query/nh_1_+_nh_1,_instant_query                                                                                    1.458m ± 2%   1.424m ±  3%        ~ (p=0.160 n=12)
Query/nh_1,_range_query_with_100_steps                                                                              831.2µ ± 3%   848.1µ ±  4%        ~ (p=0.589 n=6)
Query/nh_1_+_nh_1,_range_query_with_100_steps                                                                       1.714m ± 2%   1.653m ±  2%   -3.58% (p=0.002 n=12)
Query/nh_1,_range_query_with_1000_steps                                                                             1.611m ± 3%   1.606m ±  1%        ~ (p=0.589 n=6)
Query/nh_1_+_nh_1,_range_query_with_1000_steps                                                                      4.121m ± 1%   3.586m ±  1%  -12.97% (p=0.000 n=12)
Query/nh_100,_instant_query                                                                                         3.933m ± 1%   3.860m ±  2%   -1.84% (p=0.026 n=6)
Query/nh_100_+_nh_100,_instant_query                                                                                8.076m ± 1%   8.033m ±  1%        ~ (p=0.143 n=12)
Query/nh_100,_range_query_with_100_steps                                                                            12.22m ± 2%   12.06m ±  1%        ~ (p=0.065 n=6)
Query/nh_100_+_nh_100,_range_query_with_100_steps                                                                   39.21m ± 1%   29.53m ±  1%  -24.69% (p=0.000 n=12)
Query/nh_100,_range_query_with_1000_steps                                                                           86.65m ± 1%   86.31m ±  1%   -0.39% (p=0.041 n=6)
Query/nh_100_+_nh_100,_range_query_with_1000_steps                                                                  331.2m ± 1%   221.0m ±  1%  -33.26% (p=0.000 n=12)
Query/nh_2000,_instant_query                                                                                        63.89m ± 1%   63.88m ±  3%        ~ (p=1.000 n=6)
Query/nh_2000_+_nh_2000,_instant_query                                                                              134.3m ± 0%   131.8m ±  1%   -1.89% (p=0.000 n=12)
Query/nh_2000,_range_query_with_100_steps                                                                           224.8m ± 2%   225.8m ±  1%        ~ (p=0.394 n=6)
Query/nh_2000_+_nh_2000,_range_query_with_100_steps                                                                 836.6m ± 2%   562.5m ±  1%  -32.77% (p=0.000 n=12)
Query/nh_2000,_range_query_with_1000_steps                                                                           2.112 ± 4%    2.034 ±  3%   -3.72% (p=0.041 n=6)
Query/nh_2000_+_nh_2000,_range_query_with_1000_steps                                                                 7.968 ± 1%    4.573 ±  1%  -42.60% (p=0.000 n=12)
Query/a_1[1m],_instant_query                                                                                        686.7µ ± 4%   654.1µ ±  4%   -4.74% (p=0.041 n=6)
Query/a_100[1m],_instant_query                                                                                      1.213m ± 4%   1.136m ±  2%   -6.35% (p=0.002 n=6)
Query/a_2000[1m],_instant_query                                                                                     9.357m ± 1%   8.149m ±  1%  -12.91% (p=0.002 n=6)
Query/rate(a_1[1m]),_instant_query                                                                                  698.7µ ± 2%   673.0µ ±  5%   -3.67% (p=0.004 n=6)
Query/rate(a_1[1m]),_range_query_with_100_steps                                                                     726.2µ ± 6%   685.0µ ±  6%        ~ (p=0.180 n=6)
Query/rate(a_1[1m]),_range_query_with_1000_steps                                                                   1000.9µ ± 5%   868.4µ ±  1%  -13.24% (p=0.002 n=6)
Query/rate(a_100[1m]),_instant_query                                                                                1.319m ± 2%   1.139m ±  1%  -13.64% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_100_steps                                                                   3.423m ± 2%   2.487m ±  1%  -27.36% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_1000_steps                                                                  23.59m ± 1%   15.03m ±  2%  -36.29% (p=0.002 n=6)
Query/rate(a_2000[1m]),_instant_query                                                                              11.354m ± 1%   7.998m ±  1%  -29.56% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_100_steps                                                                  55.89m ± 1%   35.64m ±  1%  -36.22% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_1000_steps                                                                 443.7m ± 1%   276.3m ±  1%  -37.73% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_10000_steps                                                                   3.255m ± 1%   2.541m ±  1%  -21.92% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_10000_steps                                                                 231.5m ± 1%   148.3m ±  1%  -35.93% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_10000_steps                                                                 4.723 ± 2%    2.906 ±  2%  -38.48% (p=0.002 n=6)
Query/rate(nh_1[1m]),_instant_query                                                                                 713.8µ ± 3%   681.4µ ±  5%        ~ (p=0.065 n=6)
Query/rate(nh_1[1m]),_range_query_with_100_steps                                                                    902.5µ ± 3%   847.5µ ±  2%   -6.09% (p=0.002 n=6)
Query/rate(nh_1[1m]),_range_query_with_1000_steps                                                                   2.023m ± 2%   1.734m ±  1%  -14.25% (p=0.002 n=6)
Query/rate(nh_100[1m]),_instant_query                                                                               1.615m ± 2%   1.401m ±  3%  -13.23% (p=0.002 n=6)
Query/rate(nh_100[1m]),_range_query_with_100_steps                                                                  14.38m ± 1%   11.30m ±  1%  -21.43% (p=0.002 n=6)
Query/rate(nh_100[1m]),_range_query_with_1000_steps                                                                126.50m ± 1%   98.11m ±  1%  -22.44% (p=0.002 n=6)
Query/rate(nh_2000[1m]),_instant_query                                                                              17.17m ± 1%   13.23m ±  1%  -22.99% (p=0.002 n=6)
Query/rate(nh_2000[1m]),_range_query_with_100_steps                                                                 263.2m ± 1%   204.3m ±  0%  -22.37% (p=0.002 n=6)
Query/rate(nh_2000[1m]),_range_query_with_1000_steps                                                                 2.735 ± 0%    2.186 ±  1%  -20.07% (p=0.002 n=6)
Query/rate(nh_1[1m]),_range_query_with_10000_steps                                                                  14.20m ± 2%   11.14m ±  1%  -21.52% (p=0.002 n=6)
Query/rate(nh_100[1m]),_range_query_with_10000_steps                                                                 1.399 ± 1%    1.109 ±  3%  -20.69% (p=0.002 n=6)
Query/rate(nh_2000[1m]),_range_query_with_10000_steps                                                                27.61 ± 1%    21.41 ±  1%  -22.47% (p=0.002 n=6)
Query/rate(a_1[1d]),_instant_query                                                                                  2.325m ± 1%   2.048m ±  1%  -11.92% (p=0.002 n=6)
Query/rate(a_1[1d]),_range_query_with_100_steps                                                                     3.225m ± 2%   2.635m ±  1%  -18.30% (p=0.002 n=6)
Query/rate(a_1[1d]),_range_query_with_1000_steps                                                                   11.738m ± 2%   8.568m ±  2%  -27.01% (p=0.002 n=6)
Query/rate(a_100[1d]),_instant_query                                                                               108.12m ± 1%   93.68m ±  1%  -13.35% (p=0.002 n=6)
Query/rate(a_100[1d]),_range_query_with_100_steps                                                                   203.4m ± 1%   160.0m ±  1%  -21.34% (p=0.002 n=6)
Query/rate(a_100[1d]),_range_query_with_1000_steps                                                                 1052.8m ± 1%   753.9m ±  0%  -28.39% (p=0.002 n=6)
Query/rate(a_2000[1d]),_instant_query                                                                                1.954 ± 2%    1.669 ±  1%  -14.59% (p=0.002 n=6)
Query/rate(a_2000[1d]),_range_query_with_100_steps                                                                   3.840 ± 1%    2.986 ±  0%  -22.25% (p=0.002 n=6)
Query/rate(a_2000[1d]),_range_query_with_1000_steps                                                                  20.70 ± 0%    14.85 ±  0%  -28.26% (p=0.002 n=6)
Query/rate(nh_1[1h]),_instant_query                                                                                1027.4µ ± 3%   954.0µ ±  4%   -7.14% (p=0.002 n=6)
Query/rate(nh_1[1h]),_range_query_with_100_steps                                                                    1.619m ± 2%   1.265m ±  5%  -21.84% (p=0.002 n=6)
Query/rate(nh_1[1h]),_range_query_with_1000_steps                                                                   6.705m ± 1%   4.108m ±  1%  -38.73% (p=0.002 n=6)
Query/rate(nh_100[1h]),_instant_query                                                                               16.22m ± 2%   13.85m ±  0%  -14.58% (p=0.002 n=6)
Query/rate(nh_100[1h]),_range_query_with_100_steps                                                                  73.98m ± 1%   46.58m ±  1%  -37.03% (p=0.002 n=6)
Query/rate(nh_100[1h]),_range_query_with_1000_steps                                                                 605.5m ± 1%   348.9m ±  1%  -42.37% (p=0.002 n=6)
Query/rate(nh_2000[1h]),_instant_query                                                                              297.1m ± 1%   252.1m ±  1%  -15.13% (p=0.002 n=6)
Query/rate(nh_2000[1h]),_range_query_with_100_steps                                                                1475.6m ± 1%   912.2m ±  1%  -38.19% (p=0.002 n=6)
Query/rate(nh_2000[1h]),_range_query_with_1000_steps                                                                12.181 ± 0%    7.196 ±  0%  -40.93% (p=0.002 n=6)
Query/a_1_-_b_1,_instant_query                                                                                      1.324m ± 8%   1.363m ±  6%        ~ (p=0.132 n=6)
Query/a_1_-_b_1,_range_query_with_100_steps                                                                         1.469m ± 4%   1.332m ±  3%   -9.32% (p=0.002 n=6)
Query/a_1_-_b_1,_range_query_with_1000_steps                                                                        2.150m ± 2%   1.706m ±  5%  -20.66% (p=0.002 n=6)
Query/a_100_-_b_100,_instant_query                                                                                  3.794m ± 3%   3.835m ±  2%        ~ (p=0.180 n=6)
Query/a_100_-_b_100,_range_query_with_100_steps                                                                     9.681m ± 1%   5.954m ±  2%  -38.50% (p=0.002 n=6)
Query/a_100_-_b_100,_range_query_with_1000_steps                                                                    66.96m ± 1%   26.70m ±  1%  -60.12% (p=0.002 n=6)
Query/a_2000_-_b_2000,_instant_query                                                                                45.84m ± 2%   46.20m ±  1%        ~ (p=0.240 n=6)
Query/a_2000_-_b_2000,_range_query_with_100_steps                                                                  181.59m ± 1%   90.34m ±  1%  -50.25% (p=0.002 n=6)
Query/a_2000_-_b_2000,_range_query_with_1000_steps                                                                 1496.8m ± 1%   493.3m ±  1%  -67.04% (p=0.002 n=6)
Query/nh_1_/_a_1,_instant_query                                                                                     1.414m ± 6%   1.376m ±  5%        ~ (p=0.394 n=6)
Query/nh_1_/_a_1,_range_query_with_100_steps                                                                        1.627m ± 7%   1.554m ± 10%        ~ (p=0.065 n=6)
Query/nh_1_/_a_1,_range_query_with_1000_steps                                                                       3.152m ± 3%   2.716m ±  1%  -13.84% (p=0.002 n=6)
Query/nh_100_/_a_100,_instant_query                                                                                 5.898m ± 2%   5.919m ±  1%        ~ (p=0.485 n=6)
Query/nh_100_/_a_100,_range_query_with_100_steps                                                                    24.95m ± 3%   18.16m ±  1%  -27.22% (p=0.002 n=6)
Query/nh_100_/_a_100,_range_query_with_1000_steps                                                                   211.3m ± 4%   124.5m ±  2%  -41.11% (p=0.002 n=6)
Query/nh_2000_/_a_2000,_instant_query                                                                               90.16m ± 1%   89.82m ±  1%        ~ (p=0.818 n=6)
Query/nh_2000_/_a_2000,_range_query_with_100_steps                                                                  528.3m ± 1%   330.6m ±  2%  -37.41% (p=0.002 n=6)
Query/nh_2000_/_a_2000,_range_query_with_1000_steps                                                                  5.081 ± 1%    2.713 ±  1%  -46.60% (p=0.002 n=6)
Query/a_1_-_b_1,_range_query_with_10000_steps                                                                       8.821m ± 2%   4.646m ±  1%  -47.33% (p=0.002 n=6)
Query/a_100_-_b_100,_range_query_with_10000_steps                                                                   654.4m ± 1%   248.1m ±  1%  -62.08% (p=0.002 n=6)
Query/a_2000_-_b_2000,_range_query_with_10000_steps                                                                 15.520 ± 2%    4.866 ±  1%  -68.65% (p=0.002 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_instant_query                                                                   3.647m ± 2%   3.660m ±  2%        ~ (p=0.699 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_range_query_with_100_steps                                                      7.605m ± 1%   5.474m ±  1%  -28.01% (p=0.002 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_range_query_with_1000_steps                                                     46.99m ± 1%   23.20m ±  1%  -50.62% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_instant_query                                                                     35.29m ± 1%   30.05m ±  1%  -14.85% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_range_query_with_100_steps                                                       119.15m ± 1%   56.17m ±  1%  -52.86% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_range_query_with_1000_steps                                                       891.9m ± 0%   298.0m ±  1%  -66.59% (p=0.002 n=6)
Query/a_2000{l="1234"}_-_b_2000,_instant_query                                                                      23.33m ± 1%   12.78m ±  2%  -45.21% (p=0.002 n=6)
Query/a_2000{l="1234"}_-_b_2000,_range_query_with_100_steps                                                         62.96m ± 1%   19.45m ±  1%  -69.10% (p=0.002 n=6)
Query/a_2000{l="1234"}_-_b_2000,_range_query_with_1000_steps                                                       428.72m ± 1%   86.64m ±  2%  -79.79% (p=0.002 n=6)
Query/a_2000_-_b_2000{l="1234"},_instant_query                                                                      22.69m ± 1%   12.59m ±  1%  -44.51% (p=0.002 n=6)
Query/a_2000_-_b_2000{l="1234"},_range_query_with_100_steps                                                         50.42m ± 0%   19.20m ±  2%  -61.92% (p=0.002 n=6)
Query/a_2000_-_b_2000{l="1234"},_range_query_with_1000_steps                                                       312.82m ± 1%   85.19m ±  2%  -72.77% (p=0.002 n=6)
Query/sum(a_1),_instant_query                                                                                       643.6µ ± 4%   634.1µ ±  4%        ~ (p=0.310 n=6)
Query/sum(a_1),_range_query_with_100_steps                                                                          685.5µ ± 7%   656.5µ ±  2%        ~ (p=0.093 n=6)
Query/sum(a_1),_range_query_with_1000_steps                                                                         953.4µ ± 3%   873.3µ ±  9%   -8.41% (p=0.002 n=6)
Query/sum(a_100),_instant_query                                                                                     1.886m ± 2%   1.883m ±  2%        ~ (p=0.818 n=6)
Query/sum(a_100),_range_query_with_100_steps                                                                        2.993m ± 1%   2.876m ±  1%   -3.92% (p=0.002 n=6)
Query/sum(a_100),_range_query_with_1000_steps                                                                       13.99m ± 2%   12.81m ±  1%   -8.47% (p=0.002 n=6)
Query/sum(a_2000),_instant_query                                                                                    21.74m ± 1%   21.48m ±  1%        ~ (p=0.093 n=6)
Query/sum(a_2000),_range_query_with_100_steps                                                                       45.52m ± 1%   41.90m ±  2%   -7.95% (p=0.002 n=6)
Query/sum(a_2000),_range_query_with_1000_steps                                                                      282.3m ± 2%   230.9m ±  1%  -18.18% (p=0.002 n=6)
Query/sum_without_(l)(h_1),_instant_query                                                                           803.8µ ± 5%   771.2µ ±  6%   -4.06% (p=0.026 n=6)
Query/sum_without_(l)(h_1),_range_query_with_100_steps                                                              891.5µ ± 3%   854.8µ ±  3%   -4.11% (p=0.004 n=6)
…